Ver 0.1a

+++ pkgmgr.sh
@@ -136,11 +136,14 @@ if [ "$usr" == "root" -o "$usr" == "Devloc" ]; then
          if [ -f "$2" ]; then
             cp $2 $infldr/$2
             echo "$2 installed sucessfully!"
-         elif [ -f "$2.bin" ]; then
-            cp $2.bin $infldr/$2
-            echo "$2 installed sucessfully!"
-         elif [ -f "$2.sh" ]; then
-            cp $2.sh $infldr/$2
+         elif [ -f "$2.bin" -o -f "$2.sh" -o -f "$2.py" ]; then
+            if [ -f "$2.bin" ]; then
+               cp $2.bin $infldr/$2
+            elif [ -f "$2.sh" ]; then
+               cp $2.sh $infldr/$2
+            elif [ -f "$2.py" ]; then
+               cp $2.sh $infldr/$2
+            fi
             echo "$2 installed sucessfully!"
          elif [ -f "configure" -o -f "Makefile" ]; then
             if [ -f "$mkloc/$mk" -o -f "/bin/$mk" -o -f "/usr/bin/$mk" ]; then
